+++ src/lib.rs
@@ -209,6 +209,49 @@ pub fn combined_json(
 }
 
 ///
+/// Runs the disassembler for EraVM bytecode file and prints the output to stdout.
+///
+pub fn disassemble_eravm(paths: Vec<PathBuf>) -> anyhow::Result<()> {
+    let target_machine = era_compiler_llvm_context::TargetMachine::new(
+        era_compiler_common::Target::EraVM,
+        &era_compiler_llvm_context::OptimizerSettings::cycles(),
+        &[],
+    )?;
+
+    let disassemblies: Vec<(PathBuf, String)> = paths
+        .into_iter()
+        .map(|path| {
+            let bytecode = match path.extension().and_then(|extension| extension.to_str()) {
+                Some("hex") => {
+                    let string = std::fs::read_to_string(path.as_path())?;
+                    let hexadecimal_string =
+                        string.trim().strip_prefix("0x").unwrap_or(string.as_str());
+                    hex::decode(hexadecimal_string)?
+                }
+                Some("zbin") => std::fs::read(path.as_path())?,
+                Some(extension) => anyhow::bail!(
+                    "Invalid file extension: {extension}. Supported extensions: *.hex, *.zbin"
+                ),
+                None => {
+                    anyhow::bail!("Missing file extension. Supported extensions: *.hex, *.zbin")
+                }
+            };
+
+            let disassembly =
+                era_compiler_llvm_context::eravm_disassemble(&target_machine, bytecode.as_slice())?;
+            Ok((path, disassembly))
+        })
+        .collect::<anyhow::Result<Vec<(PathBuf, String)>>>()?;
+
+    for (path, disassembly) in disassemblies.into_iter() {
+        writeln!(std::io::stderr(), "File {path:?} disassembly:\n\n")?;
+        writeln!(std::io::stdout(), "{disassembly}")?;
+        writeln!(std::io::stderr(), "\n\n")?;
+    }
+    std::process::exit(era_compiler_common::EXIT_CODE_SUCCESS);
+}
+
+///
 /// Normalizes an input path by converting it to POSIX format.
 ///
 pub fn path_to_posix(path: &Path) -> anyhow::Result<PathBuf> {